support for constant variables

+++ dfsch/eqhash.h
@@ -67,7 +67,7 @@ void dfsch_eqhash_set(dfsch_eqhash_t* hash,
                       dfsch_object_t* key, dfsch_object_t* value);
 int dfsch_eqhash_unset(dfsch_eqhash_t* hash, dfsch_object_t* key);
 void dfsch_eqhash_set_flags(dfsch_eqhash_t* hash,
-                            dfsch_object_t* key, long flags);
+                            dfsch_object_t* key, short flags);
 int dfsch_eqhash_set_if_exists(dfsch_eqhash_t* hash,
                                dfsch_object_t* key, dfsch_object_t* value,
                                short* flags);
